Files app: File grid update styles
Initial re-style for file grid, updates the size, spaces and some colors
in the file grid.

Some styles are shared between file list and file grid and also previous
look'n'feel and the new one, so CSS selectors are hard to overwrite for
new look'n'feel in the file grid.

Add":not(files-ng)" to the icon selector to simplify the issue and add
"#list-container list" when the rule should apply only to file list.

Add a new test: add file-ng flag showGridViewDownloads

Test: browser_tests --gtest_filter="*showGridViewDownloads*"
Bug: 992822
Change-Id: Ia761775207377a3aa94c5fe1762702cc2afbda62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072685
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746132}