Move enum SurfaceOrigin to ui/gfx and expose to GLSurface
The purpose is to indicate where (0,0) is in an output surface. And
replace GLSurface::FlipsVertically() with a descriptive function that
returns surface origin.

It is nicer if enum SurfaceOrigin is shared between ui/gl/ and
components/viz/ so there's no logic for transforming.

There is also an instance to change in
gpu/command_buffer/common/capabilities.h. Future CL will modify it and
IPC definition that uses it.

Bug: 1051970
Change-Id: I33254620feba54a77ad51ff3fd6b82d6962d25bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083571
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Kramer Ge <fangzhoug@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746157}