media/gpu: Use STORAGE_MOJO_SHARED_MEMORY for video frames in VEA test
For trogdor, it doesn't support I420 input format, so when we run that
test it needs to use the image processor to convert the format. To make
the formats match what we use in reality, they should be either GMBs or
STORAGE_MOJO_SHARED_MEMORY. This changes the tests and the V4L2VEA so
the formats match.

BUG=b:147411004
TEST=VEA SimpleEncode test passes on trogdor w/ I420 input

Change-Id: I259c39f8c7632fb0d97d0baf1672ac05acbba6b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050895
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746212}