WebCrypto: Replace Dictionary in AlgorithmIdentifier with object
Drops a use case of Dictionary in IDL.
This CL replaces it with object, thus now AlgorithmIdentifier becomes
(object or DOMString).

As a side effect, this CL drops support of having Dictionary type
as a member of union types, and supports object type as a member of
union types instead.


Bug: 385376, 839389, 1047081
Change-Id: I24fd5813f6d7af742f46505f2b19cc2d212ec349
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066385
Reviewed-by: Eric Roman <eroman@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746225}