Add event latency to compositor latency metrics
If handling of an event leads to producing a compositor frame, an UMA
histogram is recorded measuring the time between the event timestamp and
presentation of the corresponding frame to the screen, in the form of
"EventLatency.<EventType>.TotalLatency". In future, we will add
appropriate break-down metrics, too.

Bug: 1054007

Change-Id: If85f99fb83172337c21b374073a3c2d098c2fd50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018704
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: Behdad Bakhshinategh <behdadb@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746247}