Use AppServiceAppIconLoader to load icons for ARC apps.
This is the preparing for removing ArcAppIconLoader.
AppServiceAppIconLoader is used to replace ArcAppIconLoader to load app
icons for all apps.

AppServiceAppIconLoader doesn't call OnAppImageUpdated to update for
the default icon, so remove the DCHECK.

Fix the browser tests, because AppService uses mojom, so add run loop to
to wait AppService to load the app.

BUG=10161599

Change-Id: Ifbd799911016c386fb9b3b34bfd98bffbd5d9c75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081635
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746253}