[Autofill Assistant] Move basic interactions to browser/.
This is a refactoring and should have no user-facing changes.

There are three reasons for this change:
(1) Basic interactions (i.e., interactions that are not platform-dependent) should be in a platform-agnostic location.
(2) There is now a convenient build target to unit-test those interactions.
(3) This is preparation for a follow-up CL where a new basic interaction will be added: SetChips. This will require calling Controller::SetUserActions, which
would be impossible from within platform-dependent code without exposing way too much. (side note: chips should of course just be part of the model, which would make this change trivial. Unfortunately, they currently are not, and moving them is not easy / will be done in a separate CL).

Bug: b/145043394
Change-Id: I8b26fbe15afdd8f5a1eec116c3790f67b2e63dbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074500
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#746299}