[iOS] Enabling FirstRunTestCase.testDismissFirstRun
This test has been disabled for a timeout raison. This cannot be
reproduced.
The test failed twice on ios13-beta-simulator in setup phase
which is not related to the test.
So the test can be re-enabled.

Fixed: 1055457
Change-Id: Ib51ab730bb08171fe1c57896ea558bb9ba4a0375
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083472
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746318}