[weblayer] Make sure we only use the wrapped context to inflate menus.
SelectionPopupController uses ActionMode.getMenuInflater to instantiate
a MenuInflater that uses an appropriately themed context. However this
can never work for WebLayer, because it relies on the wrapped context as
the only means of accessing its assets/resources.

WebLayer cannot rely on the current implementation because reaching the
fallback is dependent on the embedding application not having defined a
resource with the same ID.

WebView, on the other hand, must continue to use the current approach,
because it injects its assets and resources into the activity context
and can make use of activity themes, but in some cases may have to fall
back to the WebContents context because on sokme Android versions and in
some circumstances the ActionMode object has a context that has not had
webview resources injected.

Bug: 1057816
Change-Id: I417bea202ffb439fcc3960c3c5e28360b16fe9d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083476
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Commit-Queue: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746341}