Disable dialog bounds verification for ExtensionsMenuViewBrowserTest.
On Linux, the extensions menu bounds can appear off the screen in tests
on the bots, which is causing the tests to be flaky. Since the menu
doesn't have any control over where it appears, this patch just disables
those checks to stop the failures.

Bug: 1050012
Change-Id: I10bbf42cee6985eca82c60cf5c6e769cf7e81ab9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080352
Commit-Queue: Bret Sepulveda <bsep@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746357}