[serial] Ensure watchers are removed before closing port
This is a follow-up to r689238 which makes sure that the file descriptor
watchers are always removed before closing the serial port. The previous
fix handled most cases but left open the possibility that a watcher was
still registered if there wasn't a pending read or write operation to
cancel.

Bug: 1057858
Change-Id: I0c760b84cbad2ec4632c9d99e8c5b2550d770ebc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083922
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746433}