[cascade] Actually use the pending substitution cache
I forgot to actually set the CSSValue (i.e. cache key), which means the
cache was never used.

This CL adds a performance test for this which should catch regressions
in the future. The test works by specifying a couple of shorthands, each
of which has a large number of var() references that needs resolving.

Local testing on my machine shows ~170ms with the cache enabled, and
over 500ms with the cache disabled.

Change-Id: I867679969bb6617219388e0b6f28adc3ae223f5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080262
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746436}