[iOS] Adds gradient layer to sign-in migration.
Adds a gradient layer aligned with the UnifiedConsent embedded screen
that will hide part of text if the device size is small to give the
illusion that it is hidden further down on the device.

Updates the anchors of the confirmation and skip sign-in buttons.

Bug: 971989
Change-Id: I20cde343820af0d4ab5de3f45a374eaed48190b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080367
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746451}