Revert change to "actively scrolling" semantics
In https://crrev.com/2fc1cb4db60 I removed a state bit from LTHI called
|did_lock_scrolling_layer|. This bit looked like it was tracking the
same thing as "are we currently in a scroll" for which we already have
other state.

Unfortunately, this was subtly wrong on two counts:

1) |did_lock_scrolling_layer| was only set in the non-animated path;
   animated scrolls always appeared invisible to this bit.
2) This was intentionally distinguishing between "currently scrolling"
   in the sense of "we've received a ScrollBegin" and "actively
   scrolling" in the sense  of "we've received a ScrollBegin as well as
   a ScrollUpdate that's caused some scrolling".

The latter is clear from the CL that added the bit:
https://codereview.chromium.org/676853002/. The use case in mind is
pages that preventDefault all touchmoves but not the touchstart. In that
case, CurrentlyScrollingNode will return a valid node because we receive
a ScrollBegin. However, we don't want to consider the compositor in a
scroll for scheduling purposes as the page is likely handling these
touches on its own.

The effect of this is that IsActivelyScrolling now returns true for
cases like wheel scrolling. This caused a change to scheduler
prioritization and regressed checkerboarding (but improved smoothness).

This CL restores the old behavior by making use of some existing state
and also updates names and comments to make this more clear to future
readers.

Bug: 1052474
Change-Id: Ibadf6053c1cc067ec659f6b84225a78cbda437b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083941
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746478}