marketing_opt_in_screen: Add browser test and remove toggles by default
Hides marketing opt-in toggles from marketing opt-in screen - seems
weird, but the screen is also used as the final screen in the flow, with
pieces to force shelf navigation buttons in tablet mode.
Marketing opt in toggle implementation is slipping to M-83, so hide the
toggle by default (a follow-up will update the screen UI to address the
white space where the opt-in toggle should be).

Adds browser tests for the screen.

BUG=976949

Change-Id: If4c9db85be2607f99be15c9aef27f7b67fc41367
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083952
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746487}