[Paint Preview] Make DirectoryKey a class
The previous implementation of DirectoryKey used a const string making
it largely incompatible with <algorithm> which I intend to make use of
in a few places. To fix this it should be made into a class that can
be copied and swapped, but with only a constant accessor.

Bug: 1058093
Change-Id: Ib3aff42e248910bd78b9a00c98777fd670999d37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2084724
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746501}