[Autofill Assistant] Hotfix for radio buttons in FormAction.
Without this CL, UI and model can go out-of-sync when tapping the radio
button directly (see screenshots in linked bug). This is a hotfix for
this issue. A proper fix requires a refactoring of the FormAction to
make the selection state part of its model (same as was done for the
T&C selection state in the CollectUserDataModel).

Bug: b/150201921
Change-Id: Ifa0a2839c532a1862358bc68e6d0537dee93b49c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083489
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746539}