[client hints] Factor out Accept-CH parsing into common/
Needed since it sometimes need to be parsed on browser as well,
not just renderer.

This also makes it uses structured headers parsing, which is what
it is spec'd as. This changes parsing behavior somewhat; for
example accepting (and ignoring) parameters. It is also supposed
to not accept non-space whitespace (which tests are updated for),
but the parser doesn't match the spec yet.

Bug: 1050726
Change-Id: Ib3953ba74efd2ae74d9003cb59edcba9bb49010f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079295
Commit-Queue: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746540}