Discard canvas content from frame if canvas is fully repaint
In the current implementation, we always retain the content from the
previous surface. However, if the full canvas needs to be repainted in
this frame, the old content would over written. So there is no need to
retain the content from the previous frame.

The performance improvement is measure by "Blink.Canvas.IsFullRepaint".

Bug: 1054666

Change-Id: Iae766225e741439c729d75f57e30418d05fe8577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063733
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746559}