Add system row on the storage management page
The main sections on the storage management page have been added.
This change adds a "System" row so that the sizes of the storage items
of the storage page (My files, browsing data, apps etc...) add up to the
amount of "In use" space displayed at the top of the page.

This change creates a SystemSizeCalculator class that keeps track of all
calculated sizes.

The "System" storage size is displayed as
- "Calculating..." while the size of any other item is still being
calculated.
- "Unknown" if the calculation of any other row fails.
- The difference between "In use" and the sum of the other storage
items otherwise.

If the calculation of a row fails, the corresponding size in the
calculation of "System" will be 0.
In other words, when the calculation of a row fails, its size is
reflected in the System section.

Test:
unit_tests --gtest_filter="*StorageHandlerTest.SystemSize"
browser_tests --gtest_filter="*OSSettingsDevicePageTest.StorageTest"

Bug: 733192
Change-Id: Id3b82b10a2b94ed32daba085f8dce719d8272fe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060445
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746569}