Skip showing "backup successful" when the user cancels
When the user is back up their container in-advance of an upgrade the
user is shown a file-chooser. If they cancel, currently they will be
shown a "backup successful" window for 2 seconds.

We track this now by removing the 2 second timeout when the user
cancels. Technically the "backup successful" screen still flashes up
while the pre-checks run, but this is practically instant so i don't
think its worth making a more involved change.

Bug: 1052719
Change-Id: If99e5e6775ac81b6dede0c8e2a2f35b77e26fe2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079776
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#746629}