Don't apply collision detection to non-initial PIP bounds change
We adjust PIP initial bounds in exo as Android isn't aware of some
Chrome OS System UI
We should apply this logic only to the PIP initial bounds. Otherwise,
the obsolete bounds might be applied incorrectly for display
rotation, PIP fling, etc.

Whether it's entering PIP or not can be checked by whether the snap
fraction is set or not as the snap fraction is set when the initial
bounds is actually applied to the window in ClientControlledState.

BUG=b:149889040
BUG=b:149875362
TEST=Manually confirmed PIP fling and screen rotation put PIP to
the correct position.

Change-Id: I0f988b1ed6467c6d7a005bb1b0abc6e220277c5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081650
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746646}